requirements.txt
pygetwindow==0.0.9
pyautogui==0.9.54
pywin32==306
keyboard==0.13.5
opencv-python==4.9.0.80
numpy==1.26.4
mss==9.0.1
pywin32
Pillow
